+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Adapted new `ValueResolverInterface` (6.2) in all value argument resolvers
+
 ## [2.0.0] - 2022-03-12
 
 ### Added